Updated some files.
I was trying to learn how to use react routers to create multiple pages for an app, I installed the react router dependencies, added it to my project and then created a first page in whicha button allows you access the original todo app. I had to edit a ot of files and create new ones.